refactor: show warning for deprecated env
