+++ packages/phrases/src/locales/en/translation/admin-console/role-details.ts
@@ -15,7 +15,7 @@ const role_details = {
   field_description: 'Description',
   field_is_default: 'Default role',
   field_is_default_description:
-    'Set this role as a default role for new users. Multiple default roles can be set. This will also affect the default role for new users created via Management API.',
+    'Set this role as a default role for new users. Multiple default roles can be set. This will also affect the default roles for users created via Management API.',
   type_m2m_role_tag: 'Machine-to-machine app role',
   type_user_role_tag: 'User role',
   permission: {
